pyproject.toml
[project]
name = "nlp-final"
version = "0.1.0"
description = ""
authors = [
    {name = "energybubu",email = "ivanhoward011878@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click (>=8.2.1,<9.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "mistralai (>=1.7.0,<2.0.0)",
    "google-genai (>=1.16.1,<2.0.0)",
    "openai (>=1.81.0,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "torch (==2.6.0)",
    "transformers (>=4.52.2,<5.0.0)",
    "accelerate (>=1.7.0,<2.0.0)",
    "bitsandbytes (>=0.45.5,<0.46.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
